Move cyclic project loading tests into separate `project_load_cyclic.rs` file
